+++ packages/cloud-cognitive/CHANGELOG.md
@@ -3,6 +3,128 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.73.0](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.72.1...@carbon/ibm-cloud-cognitive@0.73.0) (2021-08-19)
+
+
+### Features
+
+* **CreateTearsheet:** [#1076](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/1076) create tearsheet custom step ([#1174](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/1174)) ([8d0ab04](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/8d0ab046245a9b8862e351361b08068204a97f2a))
+
+
+
+
+
+## [0.72.1](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.72.0...@carbon/ibm-cloud-cognitive@0.72.1) (2021-08-19)
+
+**Note:** Version bump only for package @carbon/ibm-cloud-cognitive
+
+
+
+
+
+# [0.72.0](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.71.8...@carbon/ibm-cloud-cognitive@0.72.0) (2021-08-19)
+
+
+### Features
+
+* update Carbon versions and dependencies ([#1168](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/1168)) ([674017f](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/674017fabc13a7737c0d16deb04ffa9872d76fe6))
+
+
+
+
+
+## [0.71.8](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.71.7...@carbon/ibm-cloud-cognitive@0.71.8) (2021-08-18)
+
+
+### Bug Fixes
+
+* **CreateTearsheet:** remove grid margin ([#1170](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/1170)) ([75bfb54](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/75bfb5462ec364d86970fc142a99b91d0c5a185b))
+
+
+
+
+
+## [0.71.7](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.71.6...@carbon/ibm-cloud-cognitive@0.71.7) (2021-08-18)
+
+**Note:** Version bump only for package @carbon/ibm-cloud-cognitive
+
+
+
+
+
+## [0.71.6](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.71.5...@carbon/ibm-cloud-cognitive@0.71.6) (2021-08-18)
+
+**Note:** Version bump only for package @carbon/ibm-cloud-cognitive
+
+
+
+
+
+## [0.71.5](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.71.4...@carbon/ibm-cloud-cognitive@0.71.5) (2021-08-18)
+
+
+### Bug Fixes
+
+* Ensure SCSS for all components and stories sets required carbon and project settings ([#1166](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/1166)) ([5c77105](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/5c77105891df498f26d3f8f6214f8d7f7fa68267))
+
+
+
+
+
+## [0.71.4](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.71.3...@carbon/ibm-cloud-cognitive@0.71.4) (2021-08-18)
+
+
+### Bug Fixes
+
+* ensure internal components don't render as canary placeholders ([#1167](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/1167)) ([0b557b6](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/0b557b6163692e796058e7d2bc3b991d019b62d9))
+
+
+
+
+
+## [0.71.3](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.71.2...@carbon/ibm-cloud-cognitive@0.71.3) (2021-08-18)
+
+**Note:** Version bump only for package @carbon/ibm-cloud-cognitive
+
+
+
+
+
+## [0.71.2](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.71.1...@carbon/ibm-cloud-cognitive@0.71.2) (2021-08-17)
+
+**Note:** Version bump only for package @carbon/ibm-cloud-cognitive
+
+
+
+
+
+## [0.71.1](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.71.0...@carbon/ibm-cloud-cognitive@0.71.1) (2021-08-17)
+
+**Note:** Version bump only for package @carbon/ibm-cloud-cognitive
+
+
+
+
+
+# [0.71.0](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.70.1...@carbon/ibm-cloud-cognitive@0.71.0) (2021-08-17)
+
+
+### Features
+
+* **toolbar:** add `Toolbar`, `ToolbarButton`, and `ToolbarGroup` components ([#1103](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/1103)) ([2d5f4cf](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/2d5f4cfb1e5c650696831a5c2c660c57e6a54f85))
+
+
+
+
+
+## [0.70.1](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.70.0...@carbon/ibm-cloud-cognitive@0.70.1) (2021-08-17)
+
+**Note:** Version bump only for package @carbon/ibm-cloud-cognitive
+
+
+
+
+
 # [0.70.0](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.69.7...@carbon/ibm-cloud-cognitive@0.70.0) (2021-08-16)